fix: fixing probability propogation
